backend/requirements.txt
3-1==1.0.0
asgiref==3.2.10
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Django==3.1
django-allauth==0.47.0
django-cors-headers==3.10.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
idna==3.3
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
stripe==2.63.0
urllib3==1.26.7
